requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
argcomplete==3.2.3
asgi-lifespan==2.1.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commitizen==3.21.3
decli==0.6.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
fastapi==0.110.1
filelock==3.13.3
frozenlist==1.4.1
gotrue==2.4.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
identify==2.5.35
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
multidict==6.0.5
nodeenv==1.8.0
openai==1.16.2
orjson==3.10.0
packaging==24.0
Pillow==10.1.0
platformdirs==4.2.0
pluggy==1.4.0
postgrest==0.16.2
pre-commit==3.7.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pytest==8.1.1
pytest-alembic==0.11.0
pytest-asyncio==0.23.6
pytest-mock==3.14.0
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
questionary==2.0.1
realtime==1.0.3
requests==2.31.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.1
supabase-py-async==2.5.6
supafunc==0.4.0
termcolor==2.4.0
tomlkit==0.12.4
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.25.1
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.18.1
